move FLRW classes
