+++ astropy/cosmology/__init__.py
@@ -8,12 +8,15 @@ detailed usage examples and references.
 
 """
 
-from . import core, funcs, units, utils
+from . import core, flrw, funcs, units, utils
 from . import io  # isort: split  # needed before 'realizations'
 from . import realizations
 from .core import *
+from .flrw import *
 from .funcs import *
 from .realizations import *
 from .utils import *
 
-__all__ = core.__all__ + realizations.__all__ + funcs.__all__ + utils.__all__
+__all__ = (core.__all__ + flrw.__all__       # cosmology classes
+           + realizations.__all__            # instances thereof
+           + funcs.__all__ + utils.__all__)  # utils
