+++ astropy/cosmology/tests/test_cosmology.py
@@ -8,51 +8,51 @@ import numpy as np
 
 import astropy.constants as const
 import astropy.units as u
-from astropy.cosmology import core, funcs, realizations
-from astropy.cosmology.realizations import WMAP5, WMAP7, WMAP9, Planck13, Planck15, Planck18
+from astropy.cosmology import Cosmology, flrw, funcs
+from astropy.cosmology.realizations import Planck13, Planck18, default_cosmology
 from astropy.units import allclose
-from astropy.utils.compat.optional_deps import HAS_SCIPY  # noqa
-from astropy.utils.exceptions import AstropyDeprecationWarning, AstropyUserWarning
+from astropy.utils.compat.optional_deps import HAS_SCIPY
+from astropy.utils.exceptions import AstropyUserWarning
 
 
 def test_init():
     """ Tests to make sure the code refuses inputs it is supposed to"""
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=-0.27)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=-0.27)
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Neff=-1)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Neff=-1)
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27,
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27,
                                    Tcmb0=u.Quantity([0.0, 2], u.K))
     with pytest.raises(ValueError):
         h0bad = u.Quantity([70, 100], u.km / u.s / u.Mpc)
-        cosmo = core.FlatLambdaCDM(H0=h0bad, Om0=0.27)
+        cosmo = flrw.FlatLambdaCDM(H0=h0bad, Om0=0.27)
     with pytest.raises(ValueError):
         bad_mnu = u.Quantity([-0.3, 0.2, 0.1], u.eV)
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.2, Tcmb0=3, m_nu=bad_mnu)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.2, Tcmb0=3, m_nu=bad_mnu)
     with pytest.raises(ValueError):
         bad_mnu = u.Quantity([0.15, 0.2, 0.1], u.eV)
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.2, Tcmb0=3, Neff=2, m_nu=bad_mnu)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.2, Tcmb0=3, Neff=2, m_nu=bad_mnu)
     with pytest.raises(ValueError):
         bad_mnu = u.Quantity([-0.3, 0.2], u.eV)  # 2, expecting 3
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.2, Tcmb0=3, m_nu=bad_mnu)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.2, Tcmb0=3, m_nu=bad_mnu)
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Ob0=-0.04)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Ob0=-0.04)
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Ob0=0.4)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Ob0=0.4)
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27)
         cosmo.Ob(1)
     with pytest.raises(ValueError):
-        cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27)
+        cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27)
         cosmo.Odm(1)
     with pytest.raises(TypeError):
-        realizations.default_cosmology.validate(4)
+        default_cosmology.validate(4)
 
 
 def test_immutability():
     """Test immutability of cosmologies."""
-    cosmo = core.Cosmology()
+    cosmo = Cosmology()
     with pytest.raises(AttributeError):
         cosmo.name = "new name"
 
@@ -63,7 +63,7 @@ def test_immutability():
 
 
 def test_basic():
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=2.0, Neff=3.04,
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=2.0, Neff=3.04,
                                Ob0=0.05, name="test", meta={"a": "b"})
     assert allclose(cosmo.Om0, 0.27)
     assert allclose(cosmo.Ode0, 0.729975, rtol=1e-4)
@@ -89,7 +89,7 @@ def test_basic():
     # Make sure setting them as quantities gives the same results
     H0 = u.Quantity(70, u.km / (u.s * u.Mpc))
     T = u.Quantity(2.0, u.K)
-    cosmo = core.FlatLambdaCDM(H0=H0, Om0=0.27, Tcmb0=T, Neff=3.04, Ob0=0.05)
+    cosmo = flrw.FlatLambdaCDM(H0=H0, Om0=0.27, Tcmb0=T, Neff=3.04, Ob0=0.05)
     assert allclose(cosmo.Om0, 0.27)
     assert allclose(cosmo.Ode0, 0.729975, rtol=1e-4)
     assert allclose(cosmo.Ob0, 0.05)
@@ -112,7 +112,7 @@ def test_basic():
 def test_units():
     """ Test if the right units are being returned"""
 
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=2.0)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=2.0)
     assert cosmo.comoving_distance(1.0).unit == u.Mpc
     assert cosmo._comoving_distance_z1z2(1.0, 2.0).unit == u.Mpc
     assert cosmo.comoving_transverse_distance(1.0).unit == u.Mpc
@@ -145,7 +145,7 @@ def test_distance_broadcast():
     """ Test array shape broadcasting for functions with single
     redshift inputs"""
 
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27,
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27,
                                m_nu=u.Quantity([0.0, 0.1, 0.011], u.eV))
     z = np.linspace(0.1, 1, 6)
     z_reshape2d = z.reshape(2, 3)
@@ -185,12 +185,12 @@ def test_distance_broadcast():
 
     # Test some dark energy models
     methods = ['Om', 'Ode', 'w', 'de_density_scale']
-    for tcosmo in [core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.5),
-                   core.wCDM(H0=70, Om0=0.27, Ode0=0.5, w0=-1.2),
-                   core.w0waCDM(H0=70, Om0=0.27, Ode0=0.5, w0=-1.2, wa=-0.2),
-                   core.wpwaCDM(H0=70, Om0=0.27, Ode0=0.5,
+    for tcosmo in [flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.5),
+                   flrw.wCDM(H0=70, Om0=0.27, Ode0=0.5, w0=-1.2),
+                   flrw.w0waCDM(H0=70, Om0=0.27, Ode0=0.5, w0=-1.2, wa=-0.2),
+                   flrw.wpwaCDM(H0=70, Om0=0.27, Ode0=0.5,
                                 wp=-1.2, wa=-0.2, zp=0.9),
-                   core.w0wzCDM(H0=70, Om0=0.27, Ode0=0.5, w0=-1.2, wz=0.1)]:
+                   flrw.w0wzCDM(H0=70, Om0=0.27, Ode0=0.5, w0=-1.2, wz=0.1)]:
         for method in methods:
             g = getattr(cosmo, method)
             value_flat = g(z)
@@ -206,7 +206,7 @@ def test_distance_broadcast():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_clone():
     """Test clone operation."""
-    cosmo = core.FlatLambdaCDM(H0=70 * u.km / u.s / u.Mpc, Om0=0.27,
+    cosmo = flrw.FlatLambdaCDM(H0=70 * u.km / u.s / u.Mpc, Om0=0.27,
                                Tcmb0=3.0 * u.K, name="test", meta={"a":"b"})
     z = np.linspace(0.1, 3, 15)
 
@@ -231,7 +231,7 @@ def test_clone():
     assert allclose(newclone.Neff, cosmo.Neff)
 
     # Compare modified version with directly instantiated one
-    cmp = core.FlatLambdaCDM(H0=60 * u.km / u.s / u.Mpc, Om0=0.27,
+    cmp = flrw.FlatLambdaCDM(H0=60 * u.km / u.s / u.Mpc, Om0=0.27,
                              Tcmb0=3.0 * u.K)
     assert newclone.__class__ == cmp.__class__
     assert cmp.name is None
@@ -267,7 +267,7 @@ def test_clone():
     assert newclone.meta == dict(a="b", zz="tops")
 
     # And direct comparison
-    cmp = core.FlatLambdaCDM(name="New name", H0=65 * u.km / u.s / u.Mpc,
+    cmp = flrw.FlatLambdaCDM(name="New name", H0=65 * u.km / u.s / u.Mpc,
                              Om0=0.27, Tcmb0=2.8 * u.K)
     assert newclone.__class__ == cmp.__class__
     assert newclone.name == cmp.name
@@ -286,7 +286,7 @@ def test_clone():
                     cmp.luminosity_distance(z))
 
     # Try a dark energy class, make sure it can handle w params
-    cosmo = core.w0waCDM(name="test w0wa", H0=70 * u.km / u.s / u.Mpc,
+    cosmo = flrw.w0waCDM(name="test w0wa", H0=70 * u.km / u.s / u.Mpc,
                          Om0=0.27, Ode0=0.5, wa=0.1, Tcmb0=4.0 * u.K)
     newclone = cosmo.clone(w0=-1.1, wa=0.2)
     assert newclone.__class__ == cosmo.__class__
@@ -316,7 +316,7 @@ def test_equality():
     assert 2 != Planck18
 
     # mismatched signatures, both directions.
-    newcosmo = core.w0waCDM(**Planck18._init_arguments, Ode0=0.6)
+    newcosmo = flrw.w0waCDM(**Planck18._init_arguments, Ode0=0.6)
     assert newcosmo != Planck18
     assert Planck18 != newcosmo
 
@@ -328,7 +328,7 @@ def test_equality():
 
 def test_xtfuncs():
     """ Test of absorption and lookback integrand"""
-    cosmo = core.LambdaCDM(70, 0.3, 0.5, Tcmb0=2.725)
+    cosmo = flrw.LambdaCDM(70, 0.3, 0.5, Tcmb0=2.725)
     z = np.array([2.0, 3.2])
     assert allclose(cosmo.lookback_time_integrand(3), 0.052218976654969378,
                     rtol=1e-4)
@@ -342,56 +342,56 @@ def test_xtfuncs():
 
 def test_repr():
     """ Test string representation of built in classes"""
-    cosmo = core.LambdaCDM(70, 0.3, 0.5, Tcmb0=2.725)
+    cosmo = flrw.LambdaCDM(70, 0.3, 0.5, Tcmb0=2.725)
     expected = ('LambdaCDM(H0=70 km / (Mpc s), Om0=0.3, '
                 'Ode0=0.5, Tcmb0=2.725 K, Neff=3.04, m_nu=[0. 0. 0.] eV, '
                 'Ob0=None)')
     assert str(cosmo) == expected
 
-    cosmo = core.LambdaCDM(70, 0.3, 0.5, Tcmb0=2.725, m_nu=u.Quantity(0.01, u.eV))
+    cosmo = flrw.LambdaCDM(70, 0.3, 0.5, Tcmb0=2.725, m_nu=u.Quantity(0.01, u.eV))
     expected = ('LambdaCDM(H0=70 km / (Mpc s), Om0=0.3, Ode0=0.5, '
                 'Tcmb0=2.725 K, Neff=3.04, m_nu=[0.01 0.01 0.01] eV, '
                 'Ob0=None)')
     assert str(cosmo) == expected
 
-    cosmo = core.FlatLambdaCDM(50.0, 0.27, Tcmb0=3, Ob0=0.05)
+    cosmo = flrw.FlatLambdaCDM(50.0, 0.27, Tcmb0=3, Ob0=0.05)
     expected = ('FlatLambdaCDM(H0=50 km / (Mpc s), Om0=0.27, '
                 'Tcmb0=3 K, Neff=3.04, m_nu=[0. 0. 0.] eV, Ob0=0.05)')
 
     assert str(cosmo) == expected
 
-    cosmo = core.wCDM(60.0, 0.27, 0.6, Tcmb0=2.725, w0=-0.8, name='test1')
+    cosmo = flrw.wCDM(60.0, 0.27, 0.6, Tcmb0=2.725, w0=-0.8, name='test1')
     expected = ('wCDM(name="test1", H0=60 km / (Mpc s), Om0=0.27, '
                 'Ode0=0.6, w0=-0.8, Tcmb0=2.725 K, Neff=3.04, '
                 'm_nu=[0. 0. 0.] eV, Ob0=None)')
     assert str(cosmo) == expected
 
-    cosmo = core.FlatwCDM(65.0, 0.27, w0=-0.6, name='test2')
+    cosmo = flrw.FlatwCDM(65.0, 0.27, w0=-0.6, name='test2')
     expected = ('FlatwCDM(name="test2", H0=65 km / (Mpc s), Om0=0.27, '
                 'w0=-0.6, Tcmb0=0 K, Neff=3.04, m_nu=None, Ob0=None)')
     assert str(cosmo) == expected
 
-    cosmo = core.w0waCDM(60.0, 0.25, 0.4, w0=-0.6, Tcmb0=2.725, wa=0.1, name='test3')
+    cosmo = flrw.w0waCDM(60.0, 0.25, 0.4, w0=-0.6, Tcmb0=2.725, wa=0.1, name='test3')
     expected = ('w0waCDM(name="test3", H0=60 km / (Mpc s), Om0=0.25, '
                 'Ode0=0.4, w0=-0.6, wa=0.1, Tcmb0=2.725 K, Neff=3.04, '
                 'm_nu=[0. 0. 0.] eV, Ob0=None)')
     assert str(cosmo) == expected
 
-    cosmo = core.Flatw0waCDM(55.0, 0.35, w0=-0.9, wa=-0.2, name='test4',
+    cosmo = flrw.Flatw0waCDM(55.0, 0.35, w0=-0.9, wa=-0.2, name='test4',
                              Ob0=0.0456789)
     expected = ('Flatw0waCDM(name="test4", H0=55 km / (Mpc s), Om0=0.35, '
                 'w0=-0.9, Tcmb0=0 K, Neff=3.04, m_nu=None, '
                 'Ob0=0.0457)')
     assert str(cosmo) == expected
 
-    cosmo = core.wpwaCDM(50.0, 0.3, 0.3, wp=-0.9, wa=-0.2,
+    cosmo = flrw.wpwaCDM(50.0, 0.3, 0.3, wp=-0.9, wa=-0.2,
                          zp=0.3, name='test5')
     expected = ('wpwaCDM(name="test5", H0=50 km / (Mpc s), Om0=0.3, '
                 'Ode0=0.3, wp=-0.9, wa=-0.2, zp=0.3, Tcmb0=0 K, '
                 'Neff=3.04, m_nu=None, Ob0=None)')
     assert str(cosmo) == expected
 
-    cosmo = core.w0wzCDM(55.0, 0.4, 0.8, w0=-1.05, wz=-0.2, Tcmb0=2.725,
+    cosmo = flrw.w0wzCDM(55.0, 0.4, 0.8, w0=-1.05, wz=-0.2, Tcmb0=2.725,
                          m_nu=u.Quantity([0.001, 0.01, 0.015], u.eV))
     expected = ('w0wzCDM(H0=55 km / (Mpc s), Om0=0.4, Ode0=0.8, w0=-1.05, '
                 'wz=-0.2 Tcmb0=2.725 K, Neff=3.04, '
@@ -404,7 +404,7 @@ def test_flat_z1():
     """ Test a flat cosmology at z=1 against several other on-line
     calculators.
     """
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=0.0)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=0.0)
     z = 1
 
     # Test values were taken from the following web cosmology
@@ -431,38 +431,35 @@ def test_flat_z1():
 def test_zeroing():
     """ Tests if setting params to 0s always respects that"""
     # Make sure Ode = 0 behaves that way
-    cosmo = core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.0)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.0)
     assert allclose(cosmo.Ode([0, 1, 2, 3]), [0, 0, 0, 0])
     assert allclose(cosmo.Ode(1), 0)
     # Ogamma0 and Onu
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=0.0)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.27, Tcmb0=0.0)
     assert allclose(cosmo.Ogamma(1.5), [0, 0, 0, 0])
     assert allclose(cosmo.Ogamma([0, 1, 2, 3]), [0, 0, 0, 0])
     assert allclose(cosmo.Onu(1.5), [0, 0, 0, 0])
     assert allclose(cosmo.Onu([0, 1, 2, 3]), [0, 0, 0, 0])
     # Obaryon
-    cosmo = core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.73, Ob0=0.0)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.73, Ob0=0.0)
     assert allclose(cosmo.Ob([0, 1, 2, 3]), [0, 0, 0, 0])
 
 
 # This class is to test whether the routines work correctly
 # if one only overloads w(z)
-class test_cos_sub(core.FLRW):
+class test_cos_sub(flrw.FLRW):
     def __init__(self):
-        core.FLRW.__init__(self, 70.0, 0.27, 0.73, Tcmb0=0.0,
-                           name="test_cos")
+        super().__init__(70.0, 0.27, 0.73, Tcmb0=0.0, name="test_cos")
         self._w0 = -0.9
 
     def w(self, z):
         return self._w0 * np.ones_like(z)
 
+
 # Similar, but with neutrinos
-
-
-class test_cos_subnu(core.FLRW):
+class test_cos_subnu(flrw.FLRW):
     def __init__(self):
-        core.FLRW.__init__(self, 70.0, 0.27, 0.73, Tcmb0=3.0,
-                           m_nu=0.1 * u.eV, name="test_cos_nu")
+        super().__init__(70.0, 0.27, 0.73, Tcmb0=3.0, m_nu=0.1 * u.eV, name="test_cos_nu")
         self._w0 = -0.8
 
     def w(self, z):
@@ -473,7 +470,7 @@ class test_cos_subnu(core.FLRW):
 def test_de_subclass():
     # This is the comparison object
     z = [0.2, 0.4, 0.6, 0.9]
-    cosmo = core.wCDM(H0=70, Om0=0.27, Ode0=0.73, w0=-0.9, Tcmb0=0.0)
+    cosmo = flrw.wCDM(H0=70, Om0=0.27, Ode0=0.73, w0=-0.9, Tcmb0=0.0)
     # Values taken from Ned Wrights advanced cosmo calculator, Aug 17 2012
     assert allclose(cosmo.luminosity_distance(z),
                     [975.5, 2158.2, 3507.3, 5773.1] * u.Mpc, rtol=1e-3)
@@ -502,7 +499,7 @@ def test_varyde_lumdist_mathematica():
 
     # w0wa models
     z = np.array([0.2, 0.4, 0.9, 1.2])
-    cosmo = core.w0waCDM(H0=70, Om0=0.2, Ode0=0.8, w0=-1.1, wa=0.2, Tcmb0=0.0)
+    cosmo = flrw.w0waCDM(H0=70, Om0=0.2, Ode0=0.8, w0=-1.1, wa=0.2, Tcmb0=0.0)
     assert allclose(cosmo.w0, -1.1)
     assert allclose(cosmo.wa, 0.2)
 
@@ -512,16 +509,16 @@ def test_varyde_lumdist_mathematica():
     assert allclose(cosmo.de_density_scale([0.0, 0.5, 1.5]),
                     [1.0, 0.9246310669529021, 0.9184087000251957])
 
-    cosmo = core.w0waCDM(H0=70, Om0=0.3, Ode0=0.7, w0=-0.9, wa=0.0, Tcmb0=0.0)
+    cosmo = flrw.w0waCDM(H0=70, Om0=0.3, Ode0=0.7, w0=-0.9, wa=0.0, Tcmb0=0.0)
     assert allclose(cosmo.luminosity_distance(z),
                     [971.667, 2141.67, 5685.96, 8107.41] * u.Mpc, rtol=1e-4)
-    cosmo = core.w0waCDM(H0=70, Om0=0.3, Ode0=0.7, w0=-0.9, wa=-0.5,
+    cosmo = flrw.w0waCDM(H0=70, Om0=0.3, Ode0=0.7, w0=-0.9, wa=-0.5,
                          Tcmb0=0.0)
     assert allclose(cosmo.luminosity_distance(z),
                     [974.087, 2157.08, 5783.92, 8274.08] * u.Mpc, rtol=1e-4)
 
     # wpwa models
-    cosmo = core.wpwaCDM(H0=70, Om0=0.2, Ode0=0.8, wp=-1.1, wa=0.2, zp=0.5,
+    cosmo = flrw.wpwaCDM(H0=70, Om0=0.2, Ode0=0.8, wp=-1.1, wa=0.2, zp=0.5,
                          Tcmb0=0.0)
     assert allclose(cosmo.wp, -1.1)
     assert allclose(cosmo.wa, 0.2)
@@ -529,7 +526,7 @@ def test_varyde_lumdist_mathematica():
     assert allclose(cosmo.luminosity_distance(z),
                     [1010.81, 2294.45, 6369.45, 9218.95] * u.Mpc, rtol=1e-4)
 
-    cosmo = core.wpwaCDM(H0=70, Om0=0.2, Ode0=0.8, wp=-1.1, wa=0.2, zp=0.9,
+    cosmo = flrw.wpwaCDM(H0=70, Om0=0.2, Ode0=0.8, wp=-1.1, wa=0.2, zp=0.9,
                          Tcmb0=0.0)
     assert allclose(cosmo.wp, -1.1)
     assert allclose(cosmo.wa, 0.2)
@@ -541,7 +538,7 @@ def test_varyde_lumdist_mathematica():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_matter():
     # Test non-relativistic matter evolution
-    tcos = core.FlatLambdaCDM(70.0, 0.3, Ob0=0.045)
+    tcos = flrw.FlatLambdaCDM(70.0, 0.3, Ob0=0.045)
     assert allclose(tcos.Om0, 0.3)
     assert allclose(tcos.H0, 70.0 * u.km / u.s / u.Mpc)
     assert allclose(tcos.Om(0), 0.3)
@@ -562,7 +559,7 @@ def test_matter():
 def test_ocurv():
     # Test Ok evolution
     # Flat, boring case
-    tcos = core.FlatLambdaCDM(70.0, 0.3)
+    tcos = flrw.FlatLambdaCDM(70.0, 0.3)
     assert allclose(tcos.Ok0, 0.0)
     assert allclose(tcos.Ok(0), 0.0)
     z = np.array([0.0, 0.5, 1.0, 2.0])
@@ -570,7 +567,7 @@ def test_ocurv():
                     rtol=1e-6)
 
     # Not flat
-    tcos = core.LambdaCDM(70.0, 0.3, 0.5, Tcmb0=u.Quantity(0.0, u.K))
+    tcos = flrw.LambdaCDM(70.0, 0.3, 0.5, Tcmb0=u.Quantity(0.0, u.K))
     assert allclose(tcos.Ok0, 0.2)
     assert allclose(tcos.Ok(0), 0.2)
     assert allclose(tcos.Ok(z), [0.2, 0.22929936, 0.21621622, 0.17307692],
@@ -584,7 +581,7 @@ def test_ocurv():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_ode():
     # Test Ode evolution, turn off neutrinos, cmb
-    tcos = core.FlatLambdaCDM(70.0, 0.3, Tcmb0=0)
+    tcos = flrw.FlatLambdaCDM(70.0, 0.3, Tcmb0=0)
     assert allclose(tcos.Ode0, 0.7)
     assert allclose(tcos.Ode(0), 0.7)
     z = np.array([0.0, 0.5, 1.0, 2.0])
@@ -609,26 +606,26 @@ def test_ogamma():
     # unfortunate, but reflects a problem with it rather than this code.
     # More accurate tests below using Mathematica
     z = np.array([1.0, 10.0, 500.0, 1000.0])
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0, Neff=3)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0, Neff=3)
     assert allclose(cosmo.angular_diameter_distance(z),
                     [1651.9, 858.2, 26.855, 13.642] * u.Mpc, rtol=5e-4)
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=2.725, Neff=3)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=2.725, Neff=3)
     assert allclose(cosmo.angular_diameter_distance(z),
                     [1651.8, 857.9, 26.767, 13.582] * u.Mpc, rtol=5e-4)
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=4.0, Neff=3)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=4.0, Neff=3)
     assert allclose(cosmo.angular_diameter_distance(z),
                     [1651.4, 856.6, 26.489, 13.405] * u.Mpc, rtol=5e-4)
 
     # Next compare with doing the integral numerically in Mathematica,
     # which allows more precision in the test.  It is at least as
     # good as 0.01%, possibly better
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0, Neff=3.04)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0, Neff=3.04)
     assert allclose(cosmo.angular_diameter_distance(z),
                     [1651.91, 858.205, 26.8586, 13.6469] * u.Mpc, rtol=1e-5)
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=2.725, Neff=3.04)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=2.725, Neff=3.04)
     assert allclose(cosmo.angular_diameter_distance(z),
                     [1651.76, 857.817, 26.7688, 13.5841] * u.Mpc, rtol=1e-5)
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=4.0, Neff=3.04)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=4.0, Neff=3.04)
     assert allclose(cosmo.angular_diameter_distance(z),
                     [1651.21, 856.411, 26.4845, 13.4028] * u.Mpc, rtol=1e-5)
 
@@ -641,7 +638,7 @@ def test_ogamma():
     Or0 = (Ogamma0h2 + Onu0h2) / 0.7 ** 2
     Om0 = 1.0 - Or0
     hubdis = (299792.458 / 70.0) * u.Mpc
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=Om0, Tcmb0=2.725, Neff=3.04)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=Om0, Tcmb0=2.725, Neff=3.04)
     targvals = 2.0 * hubdis * \
         (np.sqrt((1.0 + Or0 * z) / (1.0 + z)) - 1.0) / (Or0 - 1.0)
     assert allclose(cosmo.comoving_distance(z), targvals, rtol=1e-5)
@@ -653,7 +650,7 @@ def test_ogamma():
     # Try Tcmb0 = 4
     Or0 *= (4.0 / 2.725) ** 4
     Om0 = 1.0 - Or0
-    cosmo = core.FlatLambdaCDM(H0=70, Om0=Om0, Tcmb0=4.0, Neff=3.04)
+    cosmo = flrw.FlatLambdaCDM(H0=70, Om0=Om0, Tcmb0=4.0, Neff=3.04)
     targvals = 2.0 * hubdis * \
         (np.sqrt((1.0 + Or0 * z) / (1.0 + z)) - 1.0) / (Or0 - 1.0)
     assert allclose(cosmo.comoving_distance(z), targvals, rtol=1e-5)
@@ -661,7 +658,7 @@ def test_ogamma():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_tcmb():
-    cosmo = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=2.5)
+    cosmo = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=2.5)
     assert allclose(cosmo.Tcmb0, 2.5 * u.K)
     assert allclose(cosmo.Tcmb(2), 7.5 * u.K)
     z = [0.0, 1.0, 2.0, 3.0, 9.0]
@@ -675,7 +672,7 @@ def test_tcmb():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_tnu():
-    cosmo = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=3.0)
+    cosmo = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=3.0)
     assert allclose(cosmo.Tnu0, 2.1412975665108247 * u.K, rtol=1e-6)
     assert allclose(cosmo.Tnu(2), 6.423892699532474 * u.K, rtol=1e-6)
     z = [0.0, 1.0, 2.0, 3.0]
@@ -700,31 +697,31 @@ def test_efunc_vs_invefunc():
     # Below are the 'standard' included cosmologies
     # We do the non-standard case in test_efunc_vs_invefunc_flrw,
     # since it requires scipy
-    cosmo = core.LambdaCDM(70, 0.3, 0.5)
+    cosmo = flrw.LambdaCDM(70, 0.3, 0.5)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.LambdaCDM(70, 0.3, 0.5, m_nu=u.Quantity(0.01, u.eV))
+    cosmo = flrw.LambdaCDM(70, 0.3, 0.5, m_nu=u.Quantity(0.01, u.eV))
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.FlatLambdaCDM(50.0, 0.27)
+    cosmo = flrw.FlatLambdaCDM(50.0, 0.27)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.wCDM(60.0, 0.27, 0.6, w0=-0.8)
+    cosmo = flrw.wCDM(60.0, 0.27, 0.6, w0=-0.8)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.FlatwCDM(65.0, 0.27, w0=-0.6)
+    cosmo = flrw.FlatwCDM(65.0, 0.27, w0=-0.6)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.w0waCDM(60.0, 0.25, 0.4, w0=-0.6, wa=0.1)
+    cosmo = flrw.w0waCDM(60.0, 0.25, 0.4, w0=-0.6, wa=0.1)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.Flatw0waCDM(55.0, 0.35, w0=-0.9, wa=-0.2)
+    cosmo = flrw.Flatw0waCDM(55.0, 0.35, w0=-0.9, wa=-0.2)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.wpwaCDM(50.0, 0.3, 0.3, wp=-0.9, wa=-0.2, zp=0.3)
+    cosmo = flrw.wpwaCDM(50.0, 0.3, 0.3, wp=-0.9, wa=-0.2, zp=0.3)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
-    cosmo = core.w0wzCDM(55.0, 0.4, 0.8, w0=-1.05, wz=-0.2)
+    cosmo = flrw.w0wzCDM(55.0, 0.4, 0.8, w0=-1.05, wz=-0.2)
     assert allclose(cosmo.efunc(z0), 1.0 / cosmo.inv_efunc(z0))
     assert allclose(cosmo.efunc(z), 1.0 / cosmo.inv_efunc(z))
 
@@ -749,7 +746,7 @@ def test_efunc_vs_invefunc_flrw():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_kpc_methods():
-    cosmo = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+    cosmo = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
     assert allclose(cosmo.arcsec_per_kpc_comoving(3),
                              0.0317179167 * u.arcsec / u.kpc)
     assert allclose(cosmo.arcsec_per_kpc_proper(3),
@@ -763,9 +760,9 @@ def test_kpc_methods():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_comoving_volume():
 
-    c_flat = core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.73, Tcmb0=0.0)
-    c_open = core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.0, Tcmb0=0.0)
-    c_closed = core.LambdaCDM(H0=70, Om0=2, Ode0=0.0, Tcmb0=0.0)
+    c_flat = flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.73, Tcmb0=0.0)
+    c_open = flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.0, Tcmb0=0.0)
+    c_closed = flrw.LambdaCDM(H0=70, Om0=2, Ode0=0.0, Tcmb0=0.0)
 
     # test against ned wright's calculator (cubic Gpc)
     redshifts = np.array([0.5, 1, 2, 3, 5, 9])
@@ -789,9 +786,9 @@ def test_comoving_volume():
 def test_differential_comoving_volume():
     from scipy.integrate import quad
 
-    c_flat = core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.73, Tcmb0=0.0)
-    c_open = core.LambdaCDM(H0=70, Om0=0.27, Ode0=0.0, Tcmb0=0.0)
-    c_closed = core.LambdaCDM(H0=70, Om0=2, Ode0=0.0, Tcmb0=0.0)
+    c_flat = flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.73, Tcmb0=0.0)
+    c_open = flrw.LambdaCDM(H0=70, Om0=0.27, Ode0=0.0, Tcmb0=0.0)
+    c_closed = flrw.LambdaCDM(H0=70, Om0=2, Ode0=0.0, Tcmb0=0.0)
 
     # test that integration of differential_comoving_volume()
     #  yields same as comoving_volume()
@@ -934,7 +931,7 @@ def test_flat_open_closed_icosmo():
     dm = dm * u.Mpc
     da = da * u.Mpc
     dl = dl * u.Mpc
-    cosmo = core.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70, Tcmb0=0.0)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70, Tcmb0=0.0)
     assert allclose(cosmo.comoving_transverse_distance(redshifts), dm)
     assert allclose(cosmo.angular_diameter_distance(redshifts), da)
     assert allclose(cosmo.luminosity_distance(redshifts), dl)
@@ -943,7 +940,7 @@ def test_flat_open_closed_icosmo():
     dm = dm * u.Mpc
     da = da * u.Mpc
     dl = dl * u.Mpc
-    cosmo = core.LambdaCDM(H0=70, Om0=0.3, Ode0=0.1, Tcmb0=0.0)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.1, Tcmb0=0.0)
     assert allclose(cosmo.comoving_transverse_distance(redshifts), dm)
     assert allclose(cosmo.angular_diameter_distance(redshifts), da)
     assert allclose(cosmo.luminosity_distance(redshifts), dl)
@@ -952,7 +949,7 @@ def test_flat_open_closed_icosmo():
     dm = dm * u.Mpc
     da = da * u.Mpc
     dl = dl * u.Mpc
-    cosmo = core.LambdaCDM(H0=70, Om0=2, Ode0=0.1, Tcmb0=0.0)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=2, Ode0=0.1, Tcmb0=0.0)
     assert allclose(cosmo.comoving_transverse_distance(redshifts), dm)
     assert allclose(cosmo.angular_diameter_distance(redshifts), da)
     assert allclose(cosmo.luminosity_distance(redshifts), dl)
@@ -961,7 +958,7 @@ def test_flat_open_closed_icosmo():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_integral():
     # Test integer vs. floating point inputs
-    cosmo = core.LambdaCDM(H0=73.2, Om0=0.3, Ode0=0.50)
+    cosmo = flrw.LambdaCDM(H0=73.2, Om0=0.3, Ode0=0.50)
     assert allclose(cosmo.comoving_distance(3),
                     cosmo.comoving_distance(3.0), rtol=1e-7)
     assert allclose(cosmo.comoving_distance([1, 2, 3, 5]),
@@ -975,32 +972,32 @@ def test_integral():
 
 
 def test_wz():
-    cosmo = core.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70)
     assert allclose(cosmo.w(1.0), -1.)
     assert allclose(cosmo.w([0.1, 0.2, 0.5, 1.5, 2.5, 11.5]),
                     [-1., -1, -1, -1, -1, -1])
 
-    cosmo = core.wCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-0.5)
+    cosmo = flrw.wCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-0.5)
     assert allclose(cosmo.w(1.0), -0.5)
     assert allclose(cosmo.w([0.1, 0.2, 0.5, 1.5, 2.5, 11.5]),
                     [-0.5, -0.5, -0.5, -0.5, -0.5, -0.5])
     assert allclose(cosmo.w0, -0.5)
 
-    cosmo = core.w0wzCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wz=0.5)
+    cosmo = flrw.w0wzCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wz=0.5)
     assert allclose(cosmo.w(1.0), -0.5)
     assert allclose(cosmo.w([0.0, 0.5, 1.0, 1.5, 2.3]),
                     [-1.0, -0.75, -0.5, -0.25, 0.15])
     assert allclose(cosmo.w0, -1.0)
     assert allclose(cosmo.wz, 0.5)
 
-    cosmo = core.w0waCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wa=-0.5)
+    cosmo = flrw.w0waCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wa=-0.5)
     assert allclose(cosmo.w0, -1.0)
     assert allclose(cosmo.wa, -0.5)
     assert allclose(cosmo.w(1.0), -1.25)
     assert allclose(cosmo.w([0.0, 0.5, 1.0, 1.5, 2.3]),
                     [-1, -1.16666667, -1.25, -1.3, -1.34848485])
 
-    cosmo = core.wpwaCDM(H0=70, Om0=0.3, Ode0=0.70, wp=-0.9,
+    cosmo = flrw.wpwaCDM(H0=70, Om0=0.3, Ode0=0.70, wp=-0.9,
                          wa=0.2, zp=0.5)
     assert allclose(cosmo.wp, -0.9)
     assert allclose(cosmo.wa, 0.2)
@@ -1013,7 +1010,7 @@ def test_wz():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_de_densityscale():
-    cosmo = core.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70)
+    cosmo = flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70)
     z = np.array([0.1, 0.2, 0.5, 1.5, 2.5])
     assert allclose(cosmo.de_density_scale(z),
                     [1.0, 1.0, 1.0, 1.0, 1.0])
@@ -1023,7 +1020,7 @@ def test_de_densityscale():
     assert allclose(cosmo.de_density_scale([1, 2, 3]),
                        cosmo.de_density_scale([1., 2., 3.]), rtol=1e-7)
 
-    cosmo = core.wCDM(H0=70, Om0=0.3, Ode0=0.60, w0=-0.5)
+    cosmo = flrw.wCDM(H0=70, Om0=0.3, Ode0=0.60, w0=-0.5)
     assert allclose(cosmo.de_density_scale(z),
                     [1.15369, 1.31453, 1.83712, 3.95285, 6.5479],
                     rtol=1e-4)
@@ -1032,7 +1029,7 @@ def test_de_densityscale():
     assert allclose(cosmo.de_density_scale([1, 2, 3]),
                     cosmo.de_density_scale([1., 2., 3.]), rtol=1e-7)
 
-    cosmo = core.w0wzCDM(H0=70, Om0=0.3, Ode0=0.50, w0=-1, wz=0.5)
+    cosmo = flrw.w0wzCDM(H0=70, Om0=0.3, Ode0=0.50, w0=-1, wz=0.5)
     assert allclose(cosmo.de_density_scale(z),
                     [0.746048, 0.5635595, 0.25712378, 0.026664129,
                      0.0035916468], rtol=1e-4)
@@ -1041,7 +1038,7 @@ def test_de_densityscale():
     assert allclose(cosmo.de_density_scale([1, 2, 3]),
                     cosmo.de_density_scale([1., 2., 3.]), rtol=1e-7)
 
-    cosmo = core.w0waCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wa=-0.5)
+    cosmo = flrw.w0waCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wa=-0.5)
     assert allclose(cosmo.de_density_scale(z),
                     [0.9934201, 0.9767912, 0.897450,
                      0.622236, 0.4458753], rtol=1e-4)
@@ -1050,7 +1047,7 @@ def test_de_densityscale():
     assert allclose(cosmo.de_density_scale([1, 2, 3]),
                        cosmo.de_density_scale([1., 2., 3.]), rtol=1e-7)
 
-    cosmo = core.wpwaCDM(H0=70, Om0=0.3, Ode0=0.70, wp=-0.9,
+    cosmo = flrw.wpwaCDM(H0=70, Om0=0.3, Ode0=0.70, wp=-0.9,
                          wa=0.2, zp=0.5)
     assert allclose(cosmo.de_density_scale(z),
                     [1.012246048, 1.0280102, 1.087439,
@@ -1064,7 +1061,7 @@ def test_de_densityscale():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_age():
     # WMAP7 but with Omega_relativisitic = 0
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
     assert allclose(tcos.hubble_time, 13.889094057856937 * u.Gyr)
     assert allclose(tcos.age(4), 1.5823603508870991 * u.Gyr)
     assert allclose(tcos.age([1., 5.]),
@@ -1072,12 +1069,12 @@ def test_age():
     assert allclose(tcos.age([1, 5]), [5.97113193, 1.20553129] * u.Gyr)
 
     # Add relativistic species
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=3.0)
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=3.0)
     assert allclose(tcos.age(4), 1.5773003779230699 * u.Gyr)
     assert allclose(tcos.age([1, 5]), [5.96344942, 1.20093077] * u.Gyr)
 
     # And massive neutrinos
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=3.0,
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=3.0,
                               m_nu=0.1 * u.eV)
     assert allclose(tcos.age(4), 1.5546485439853412 * u.Gyr)
     assert allclose(tcos.age([1, 5]), [5.88448152, 1.18383759] * u.Gyr)
@@ -1086,7 +1083,7 @@ def test_age():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_distmod():
     # WMAP7 but with Omega_relativisitic = 0
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
     assert allclose(tcos.hubble_distance, 4258.415596590909 * u.Mpc)
     assert allclose(tcos.distmod([1, 5]),
                     [44.124857, 48.40167258] * u.mag)
@@ -1098,7 +1095,7 @@ def test_distmod():
 def test_neg_distmod():
     # Cosmology with negative luminosity distances (perfectly okay,
     #  if obscure)
-    tcos = core.LambdaCDM(70, 0.2, 1.3, Tcmb0=0)
+    tcos = flrw.LambdaCDM(70, 0.2, 1.3, Tcmb0=0)
     assert allclose(tcos.luminosity_distance([50, 100]),
                     [16612.44047622, -46890.79092244] * u.Mpc)
     assert allclose(tcos.distmod([50, 100]),
@@ -1113,7 +1110,7 @@ def test_critical_density():
     # These tests will fail if astropy.const starts returning non-mks
     #  units by default; see the comment at the top of core.py.
     # critical_density0 is inversely proportional to G.
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
     fac = (const.G / codata2014.G).to(u.dimensionless_unscaled).value
     assert allclose(tcos.critical_density0 * fac,
                     9.309668456020899e-30 * (u.g / u.cm**3))
@@ -1129,7 +1126,7 @@ def test_critical_density():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_comoving_distance_z1z2():
-    tcos = core.LambdaCDM(100, 0.3, 0.8, Tcmb0=0.0)
+    tcos = flrw.LambdaCDM(100, 0.3, 0.8, Tcmb0=0.0)
     with pytest.raises(ValueError):  # test diff size z1, z2 fail
         tcos._comoving_distance_z1z2((1, 2), (3, 4, 5))
     # Comoving distances are invertible
@@ -1154,13 +1151,13 @@ def test_age_in_special_cosmologies():
 
     Some analytic solutions fail at these critical points.
     """
-    c_dS = core.FlatLambdaCDM(100, 0, Tcmb0=0)
+    c_dS = flrw.FlatLambdaCDM(100, 0, Tcmb0=0)
     assert allclose(c_dS.age(z=0), np.inf * u.Gyr)
     assert allclose(c_dS.age(z=1), np.inf * u.Gyr)
     assert allclose(c_dS.lookback_time(z=0), 0 * u.Gyr)
     assert allclose(c_dS.lookback_time(z=1), 6.777539216261741 * u.Gyr)
 
-    c_EdS = core.FlatLambdaCDM(100, 1, Tcmb0=0)
+    c_EdS = flrw.FlatLambdaCDM(100, 1, Tcmb0=0)
     assert allclose(c_EdS.age(z=0), 6.518614811154189 * u.Gyr)
     assert allclose(c_EdS.age(z=1), 2.3046783684542738 * u.Gyr)
     assert allclose(c_EdS.lookback_time(z=0), 0 * u.Gyr)
@@ -1173,26 +1170,26 @@ def test_distance_in_special_cosmologies():
 
     Some analytic solutions fail at these critical points.
     """
-    c_dS = core.FlatLambdaCDM(100, 0, Tcmb0=0)
+    c_dS = flrw.FlatLambdaCDM(100, 0, Tcmb0=0)
     assert allclose(c_dS.comoving_distance(z=0), 0 * u.Mpc)
     assert allclose(c_dS.comoving_distance(z=1), 2997.92458 * u.Mpc)
 
-    c_EdS = core.FlatLambdaCDM(100, 1, Tcmb0=0)
+    c_EdS = flrw.FlatLambdaCDM(100, 1, Tcmb0=0)
     assert allclose(c_EdS.comoving_distance(z=0), 0 * u.Mpc)
     assert allclose(c_EdS.comoving_distance(z=1), 1756.1435599923348 * u.Mpc)
 
-    c_dS = core.LambdaCDM(100, 0, 1, Tcmb0=0)
+    c_dS = flrw.LambdaCDM(100, 0, 1, Tcmb0=0)
     assert allclose(c_dS.comoving_distance(z=0), 0 * u.Mpc)
     assert allclose(c_dS.comoving_distance(z=1), 2997.92458 * u.Mpc)
 
-    c_EdS = core.LambdaCDM(100, 1, 0, Tcmb0=0)
+    c_EdS = flrw.LambdaCDM(100, 1, 0, Tcmb0=0)
     assert allclose(c_EdS.comoving_distance(z=0), 0 * u.Mpc)
     assert allclose(c_EdS.comoving_distance(z=1), 1756.1435599923348 * u.Mpc)
 
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_comoving_transverse_distance_z1z2():
-    tcos = core.FlatLambdaCDM(100, 0.3, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(100, 0.3, Tcmb0=0.0)
     with pytest.raises(ValueError):  # test diff size z1, z2 fail
         tcos._comoving_transverse_distance_z1z2((1, 2), (3, 4, 5))
     # Tests that should actually work, target values computed with
@@ -1210,7 +1207,7 @@ def test_comoving_transverse_distance_z1z2():
                     tcos._comoving_transverse_distance_z1z2(z1, z2))
 
     # Test Flat Universe with Omega_M > 1.  Rarely used, but perfectly valid.
-    tcos = core.FlatLambdaCDM(100, 1.5, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(100, 1.5, Tcmb0=0.0)
     results = (2202.72682564,
                1559.51679971,
                -643.21002593,
@@ -1229,7 +1226,7 @@ def test_comoving_transverse_distance_z1z2():
                     tcos._comoving_transverse_distance_z1z2(z1, z2))
     # Test non-flat cases to avoid simply testing
     # comoving_distance_z1z2. Test array, array case.
-    tcos = core.LambdaCDM(100, 0.3, 0.5, Tcmb0=0.0)
+    tcos = flrw.LambdaCDM(100, 0.3, 0.5, Tcmb0=0.0)
     results = (3535.931375645655,
                2226.430046551708,
                -1208.6817970036532,
@@ -1240,7 +1237,7 @@ def test_comoving_transverse_distance_z1z2():
                     results)
 
     # Test positive curvature with scalar, array combination.
-    tcos = core.LambdaCDM(100, 1.0, 0.2, Tcmb0=0.0)
+    tcos = flrw.LambdaCDM(100, 1.0, 0.2, Tcmb0=0.0)
     z1 = 0.1
     z2 = 0, 0.1, 0.2, 0.5, 1.1, 2
     results = (-281.31602666724865,
@@ -1256,7 +1253,7 @@ def test_comoving_transverse_distance_z1z2():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_angular_diameter_distance_z1z2():
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
     with pytest.raises(ValueError):  # test diff size z1, z2 fail
         tcos.angular_diameter_distance_z1z2([1, 2], [3, 4, 5])
     # Tests that should actually work
@@ -1290,18 +1287,18 @@ def test_angular_diameter_distance_z1z2():
                     results)
 
     # Non-flat (positive Ok0) test
-    tcos = core.LambdaCDM(H0=70.4, Om0=0.2, Ode0=0.5, Tcmb0=0.0)
+    tcos = flrw.LambdaCDM(H0=70.4, Om0=0.2, Ode0=0.5, Tcmb0=0.0)
     assert allclose(tcos.angular_diameter_distance_z1z2(1, 2),
                     620.1175337852428 * u.Mpc)
     # Non-flat (negative Ok0) test
-    tcos = core.LambdaCDM(H0=100, Om0=2, Ode0=1, Tcmb0=0.0)
+    tcos = flrw.LambdaCDM(H0=100, Om0=2, Ode0=1, Tcmb0=0.0)
     assert allclose(tcos.angular_diameter_distance_z1z2(1, 2),
                     228.42914659246014 * u.Mpc)
 
 
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_absorption_distance():
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
     assert allclose(tcos.absorption_distance([1, 3]),
                     [1.72576635, 7.98685853])
     assert allclose(tcos.absorption_distance([1., 3.]),
@@ -1313,7 +1310,7 @@ def test_absorption_distance():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_massivenu_basic():
     # Test no neutrinos case
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Neff=4.05,
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Neff=4.05,
                               Tcmb0=2.725 * u.K, m_nu=0)
     assert allclose(tcos.Neff, 4.05)
     assert not tcos.has_massive_nu
@@ -1327,13 +1324,13 @@ def test_massivenu_basic():
                     rtol=1e-6)
 
     # Alternative no neutrinos case
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=0 * u.K,
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=0 * u.K,
                               m_nu=str((0.4 * u.eV).to(u.g, u.mass_energy())))
     assert not tcos.has_massive_nu
     assert tcos.m_nu is None
 
     # Test basic setting, retrieval of values
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=2.725 * u.K,
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=2.725 * u.K,
                               m_nu=u.Quantity([0.0, 0.01, 0.02], u.eV))
     assert tcos.has_massive_nu
     mnu = tcos.m_nu
@@ -1342,7 +1339,7 @@ def test_massivenu_basic():
     assert allclose(mnu, [0.0, 0.01, 0.02] * u.eV)
 
     # All massive neutrinos case
-    tcos = core.FlatLambdaCDM(70.4, 0.272, Tcmb0=2.725,
+    tcos = flrw.FlatLambdaCDM(70.4, 0.272, Tcmb0=2.725,
                               m_nu=u.Quantity(0.1, u.eV), Neff=3.1)
     assert allclose(tcos.Neff, 3.1)
     assert tcos.has_massive_nu
@@ -1364,122 +1361,122 @@ def test_distances():
 
     # The pattern here is: no relativistic species, the relativistic
     # species with massless neutrinos, then massive neutrinos
-    cos = core.LambdaCDM(75.0, 0.25, 0.5, Tcmb0=0.0)
+    cos = flrw.LambdaCDM(75.0, 0.25, 0.5, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [2953.93001902, 4616.7134253, 5685.07765971,
                      6440.80611897] * u.Mpc, rtol=1e-4)
-    cos = core.LambdaCDM(75.0, 0.25, 0.6, Tcmb0=3.0, Neff=3,
+    cos = flrw.LambdaCDM(75.0, 0.25, 0.6, Tcmb0=3.0, Neff=3,
                          m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [3037.12620424, 4776.86236327, 5889.55164479,
                      6671.85418235] * u.Mpc, rtol=1e-4)
-    cos = core.LambdaCDM(75.0, 0.3, 0.4, Tcmb0=3.0, Neff=3,
+    cos = flrw.LambdaCDM(75.0, 0.3, 0.4, Tcmb0=3.0, Neff=3,
                          m_nu=u.Quantity(10.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2471.80626824, 3567.1902565, 4207.15995626,
                      4638.20476018] * u.Mpc, rtol=1e-4)
     # Flat
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=0.0)
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [3180.83488552, 5060.82054204, 6253.6721173,
                      7083.5374303] * u.Mpc, rtol=1e-4)
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
                               m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [3180.42662867, 5059.60529655, 6251.62766102,
                      7080.71698117] * u.Mpc, rtol=1e-4)
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
                               m_nu=u.Quantity(10.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2337.54183142, 3371.91131264, 3988.40711188,
                      4409.09346922] * u.Mpc, rtol=1e-4)
     # Add w
-    cos = core.FlatwCDM(75.0, 0.25, w0=-1.05, Tcmb0=0.0)
+    cos = flrw.FlatwCDM(75.0, 0.25, w0=-1.05, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [3216.8296894, 5117.2097601, 6317.05995437,
                      7149.68648536] * u.Mpc, rtol=1e-4)
-    cos = core.FlatwCDM(75.0, 0.25, w0=-0.95, Tcmb0=3.0, Neff=3,
+    cos = flrw.FlatwCDM(75.0, 0.25, w0=-0.95, Tcmb0=3.0, Neff=3,
                     m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [3143.56537758, 5000.32196494, 6184.11444601,
                      7009.80166062] * u.Mpc, rtol=1e-4)
-    cos = core.FlatwCDM(75.0, 0.25, w0=-0.9, Tcmb0=3.0, Neff=3,
+    cos = flrw.FlatwCDM(75.0, 0.25, w0=-0.9, Tcmb0=3.0, Neff=3,
                     m_nu=u.Quantity(10.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2337.76035371, 3372.1971387, 3988.71362289,
                      4409.40817174] * u.Mpc, rtol=1e-4)
     # Non-flat w
-    cos = core.wCDM(75.0, 0.25, 0.4, w0=-0.9, Tcmb0=0.0)
+    cos = flrw.wCDM(75.0, 0.25, 0.4, w0=-0.9, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [2849.6163356, 4428.71661565, 5450.97862778,
                      6179.37072324] * u.Mpc, rtol=1e-4)
-    cos = core.wCDM(75.0, 0.25, 0.4, w0=-1.1, Tcmb0=3.0, Neff=3,
+    cos = flrw.wCDM(75.0, 0.25, 0.4, w0=-1.1, Tcmb0=3.0, Neff=3,
                     m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2904.35580229, 4511.11471267, 5543.43643353,
                      6275.9206788] * u.Mpc, rtol=1e-4)
-    cos = core.wCDM(75.0, 0.25, 0.4, w0=-0.9, Tcmb0=3.0, Neff=3,
+    cos = flrw.wCDM(75.0, 0.25, 0.4, w0=-0.9, Tcmb0=3.0, Neff=3,
                     m_nu=u.Quantity(10.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2473.32522734, 3581.54519631, 4232.41674426,
                      4671.83818117] * u.Mpc, rtol=1e-4)
     # w0wa
-    cos = core.w0waCDM(75.0, 0.3, 0.6, w0=-0.9, wa=0.1, Tcmb0=0.0)
+    cos = flrw.w0waCDM(75.0, 0.3, 0.6, w0=-0.9, wa=0.1, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [2937.7807638, 4572.59950903, 5611.52821924,
                      6339.8549956] * u.Mpc, rtol=1e-4)
-    cos = core.w0waCDM(75.0, 0.25, 0.5, w0=-0.9, wa=0.1, Tcmb0=3.0, Neff=3,
+    cos = flrw.w0waCDM(75.0, 0.25, 0.5, w0=-0.9, wa=0.1, Tcmb0=3.0, Neff=3,
                        m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2907.34722624, 4539.01723198, 5593.51611281,
                      6342.3228444] * u.Mpc, rtol=1e-4)
-    cos = core.w0waCDM(75.0, 0.25, 0.5, w0=-0.9, wa=0.1, Tcmb0=3.0, Neff=3,
+    cos = flrw.w0waCDM(75.0, 0.25, 0.5, w0=-0.9, wa=0.1, Tcmb0=3.0, Neff=3,
                        m_nu=u.Quantity(10.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2507.18336722, 3633.33231695, 4292.44746919,
                      4736.35404638] * u.Mpc, rtol=1e-4)
     # Flatw0wa
-    cos = core.Flatw0waCDM(75.0, 0.25, w0=-0.95, wa=0.15, Tcmb0=0.0)
+    cos = flrw.Flatw0waCDM(75.0, 0.25, w0=-0.95, wa=0.15, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [3123.29892781, 4956.15204302, 6128.15563818,
                      6948.26480378] * u.Mpc, rtol=1e-4)
-    cos = core.Flatw0waCDM(75.0, 0.25, w0=-0.95, wa=0.15, Tcmb0=3.0, Neff=3,
+    cos = flrw.Flatw0waCDM(75.0, 0.25, w0=-0.95, wa=0.15, Tcmb0=3.0, Neff=3,
                            m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [3122.92671907, 4955.03768936, 6126.25719576,
                      6945.61856513] * u.Mpc, rtol=1e-4)
-    cos = core.Flatw0waCDM(75.0, 0.25, w0=-0.95, wa=0.15, Tcmb0=3.0, Neff=3,
+    cos = flrw.Flatw0waCDM(75.0, 0.25, w0=-0.95, wa=0.15, Tcmb0=3.0, Neff=3,
                            m_nu=u.Quantity(10.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2337.70072701, 3372.13719963, 3988.6571093,
                      4409.35399673] * u.Mpc, rtol=1e-4)
     # wpwa
-    cos = core.wpwaCDM(75.0, 0.3, 0.6, wp=-0.9, zp=0.5, wa=0.1, Tcmb0=0.0)
+    cos = flrw.wpwaCDM(75.0, 0.3, 0.6, wp=-0.9, zp=0.5, wa=0.1, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [2954.68975298, 4599.83254834, 5643.04013201,
                      6373.36147627] * u.Mpc, rtol=1e-4)
-    cos = core.wpwaCDM(75.0, 0.25, 0.5, wp=-0.9, zp=0.4, wa=0.1,
+    cos = flrw.wpwaCDM(75.0, 0.25, 0.5, wp=-0.9, zp=0.4, wa=0.1,
                        Tcmb0=3.0, Neff=3, m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2919.00656215, 4558.0218123, 5615.73412391,
                      6366.10224229] * u.Mpc, rtol=1e-4)
-    cos = core.wpwaCDM(75.0, 0.25, 0.5, wp=-0.9, zp=1.0, wa=0.1, Tcmb0=3.0,
+    cos = flrw.wpwaCDM(75.0, 0.25, 0.5, wp=-0.9, zp=1.0, wa=0.1, Tcmb0=3.0,
                        Neff=4, m_nu=u.Quantity(5.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2629.48489827, 3874.13392319, 4614.31562397,
                      5116.51184842] * u.Mpc, rtol=1e-4)
 
     # w0wz
-    cos = core.w0wzCDM(75.0, 0.3, 0.6, w0=-0.9, wz=0.1, Tcmb0=0.0)
+    cos = flrw.w0wzCDM(75.0, 0.3, 0.6, w0=-0.9, wz=0.1, Tcmb0=0.0)
     assert allclose(cos.comoving_distance(z),
                     [3051.68786716, 4756.17714818, 5822.38084257,
                      6562.70873734] * u.Mpc, rtol=1e-4)
-    cos = core.w0wzCDM(75.0, 0.25, 0.5, w0=-0.9, wz=0.1,
+    cos = flrw.w0wzCDM(75.0, 0.25, 0.5, w0=-0.9, wz=0.1,
                        Tcmb0=3.0, Neff=3, m_nu=u.Quantity(0.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2997.8115653, 4686.45599916, 5764.54388557,
                      6524.17408738] * u.Mpc, rtol=1e-4)
-    cos = core.w0wzCDM(75.0, 0.25, 0.5, w0=-0.9, wz=0.1, Tcmb0=3.0,
+    cos = flrw.w0wzCDM(75.0, 0.25, 0.5, w0=-0.9, wz=0.1, Tcmb0=3.0,
                        Neff=4, m_nu=u.Quantity(5.0, u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2676.73467639, 3940.57967585, 4686.90810278,
@@ -1488,22 +1485,22 @@ def test_distances():
     # Also test different numbers of massive neutrinos
     # for FlatLambdaCDM to give the scalar nu density functions a
     # work out
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0,
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0,
                              m_nu=u.Quantity([10.0, 0, 0], u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2777.71589173, 4186.91111666, 5046.0300719,
                      5636.10397302] * u.Mpc, rtol=1e-4)
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0,
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0,
                              m_nu=u.Quantity([10.0, 5, 0], u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2636.48149391, 3913.14102091, 4684.59108974,
                      5213.07557084] * u.Mpc, rtol=1e-4)
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0,
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0,
                              m_nu=u.Quantity([4.0, 5, 9], u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2563.5093049, 3776.63362071, 4506.83448243,
                      5006.50158829] * u.Mpc, rtol=1e-4)
-    cos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=4.2,
+    cos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=4.2,
                              m_nu=u.Quantity([1.0, 4.0, 5, 9], u.eV))
     assert allclose(cos.comoving_distance(z),
                     [2525.58017482, 3706.87633298, 4416.58398847,
@@ -1523,7 +1520,7 @@ def test_massivenu_density():
     nuprefac = 7.0 / 8.0 * (4.0 / 11.0) ** (4.0 / 3.0)
     #  First try 3 massive neutrinos, all 100 eV -- note this is a universe
     #  seriously dominated by neutrinos!
-    tcos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
+    tcos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
                               m_nu=u.Quantity(100.0, u.eV))
     assert tcos.has_massive_nu
     assert tcos.Neff == 3
@@ -1533,7 +1530,7 @@ def test_massivenu_density():
     assert allclose(tcos.efunc([0.0, 1.0]), [1.0, 7.46144727668], rtol=5e-3)
 
     # Next, slightly less massive
-    tcos = core.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
+    tcos = flrw.FlatLambdaCDM(75.0, 0.25, Tcmb0=3.0, Neff=3,
                               m_nu=u.Quantity(0.25, u.eV))
     nurel_exp = nuprefac * tcos.Neff * np.array([429.924, 214.964, 143.312,
                                                  39.1005, 1.11086])
@@ -1546,7 +1543,7 @@ def test_massivenu_density():
     assert allclose(tcos.Onu(ztest), onu_exp, rtol=5e-3)
 
     # And fairly light
-    tcos = core.FlatLambdaCDM(80.0, 0.30, Tcmb0=3.0, Neff=3,
+    tcos = flrw.FlatLambdaCDM(80.0, 0.30, Tcmb0=3.0, Neff=3,
                               m_nu=u.Quantity(0.01, u.eV))
 
     nurel_exp = nuprefac * tcos.Neff * np.array([17.2347, 8.67345, 5.84348,
@@ -1562,7 +1559,7 @@ def test_massivenu_density():
                     rtol=1e-4)
 
     # Now a mixture of neutrino masses, with non-integer Neff
-    tcos = core.FlatLambdaCDM(80.0, 0.30, Tcmb0=3.0, Neff=3.04,
+    tcos = flrw.FlatLambdaCDM(80.0, 0.30, Tcmb0=3.0, Neff=3.04,
                               m_nu=u.Quantity([0.0, 0.01, 0.25], u.eV))
     nurel_exp = nuprefac * tcos.Neff * \
                 np.array([149.386233, 74.87915, 50.0518,
@@ -1583,7 +1580,7 @@ def test_massivenu_density():
 @pytest.mark.skipif('not HAS_SCIPY')
 def test_elliptic_comoving_distance_z1z2():
     """Regression test for #8388."""
-    cosmo = core.LambdaCDM(70., 2.3, 0.05, Tcmb0=0)
+    cosmo = flrw.LambdaCDM(70., 2.3, 0.05, Tcmb0=0)
     z = 0.2
     assert allclose(cosmo.comoving_distance(z),
                     cosmo._integral_comoving_distance_z1z2(0., z))
@@ -1592,10 +1589,10 @@ def test_elliptic_comoving_distance_z1z2():
 
 
 SPECIALIZED_COMOVING_DISTANCE_COSMOLOGIES = [
-    core.FlatLambdaCDM(H0=70, Om0=0.0, Tcmb0=0.0),  # de Sitter
-    core.FlatLambdaCDM(H0=70, Om0=1.0, Tcmb0=0.0),  # Einstein - de Sitter
-    core.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0.0),  # Hypergeometric
-    core.LambdaCDM(H0=70, Om0=0.3, Ode0=0.6, Tcmb0=0.0),  # Elliptic
+    flrw.FlatLambdaCDM(H0=70, Om0=0.0, Tcmb0=0.0),  # de Sitter
+    flrw.FlatLambdaCDM(H0=70, Om0=1.0, Tcmb0=0.0),  # Einstein - de Sitter
+    flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0.0),  # Hypergeometric
+    flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.6, Tcmb0=0.0),  # Elliptic
 ]
