+++ docs/whatsnew/5.0.rst
@@ -39,7 +39,7 @@ e.g. enabling::
     >>> from astropy.cosmology import Planck18
     >>> cm = Planck18.to_format("mapping")
     >>> cm
-    {'cosmology': <class 'astropy.cosmology.core.FlatLambdaCDM'>,
+    {'cosmology': <class 'astropy.cosmology.flrw.FlatLambdaCDM'>,
      'name': 'Planck18',
      'H0': <Quantity 67.66 km / (Mpc s)>,
      'Om0': 0.30966,
